+++ knf/knf_test.go
@@ -2,8 +2,8 @@ package knf
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 //                                                                                    //
-//                     Copyright (c) 2009-2016 Essential Kaos                         //
-//      Essential Kaos Open Source License <http://essentialkaos.com/ekol?en>         //
+//                     Copyright (c) 2009-2017 ESSENTIAL KAOS                         //
+//        Essential Kaos Open Source License <https://essentialkaos.com/ekol>         //
 //                                                                                    //
 // ////////////////////////////////////////////////////////////////////////////////// //
 
@@ -23,7 +23,7 @@ const _CONFIG_DATA = `
 test1:      1
             test2:2
 
-      test3: 3
+		test3: 3
 
 [string]
   test1: test