+++ changelog.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+#### v6.1.0
+
+* `[knf]` Added tabs support in indentation
+* `[timeutil]` Added new sequences `%n` (_new line symbol_) and `%K` (_milliseconds_)
+* `[timeutil]` Code refactoring
+
 #### v6.0.0
 
 * `[passwd]` Much secure hash generation (now with sha512, bcrypt and AES)